fix: change temperature from BigDecimal to Double

+++ src/main/java/com/damvih/dto/WeatherViewDto.java
@@ -18,8 +18,8 @@ public class WeatherViewDto {
     private String locationName;
     private String country;
 
-    private BigDecimal temperature;
-    private BigDecimal feelsLike;
+    private Double temperature;
+    private Double feelsLike;
     private String description;
 
     private Integer windSpeed;
